# services/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Fitness Planner API
  summary: Cloud API for synchronized personal training plans and workout records
  description: Server-authoritative exercise catalog and immutable training plan versions, with offline-friendly idempotent
    workout synchronization.
  version: 1.0.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - authentication
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - authentication
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - authentication
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/LogoutRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/me:
    get:
      tags:
      - authentication
      summary: Me
      operationId: me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
...
  /api/v1/cardio-sessions:
    post:
      tags:
      - cardio sessions
      summary: Create Cardio Session
      operationId: create_cardio_session_api_v1_cardio_sessions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardioSessionUpsert'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardioSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - cardio sessions
      summary: List Cardio Sessions
      operationId: list_cardio_sessions_api_v1_cardio_sessions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: local_date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Local Date From
      - name: local_date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Local Date To
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Deleted
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardioSessionPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CardioSessionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        client_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Id
        source:
          type: string
          title: Source
          default: android
        source_device:
          type: string
          title: Source Device
          default: android
        client_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Version
        local_date:
          type: string
          format: date
          title: Local Date
        activity:
          type: string
          title: Activity
        activity_type:
          type: string
          title: Activity Type
        duration_minutes:
          type: integer
          title: Duration Minutes
        duration_seconds:
          type: integer
          title: Duration Seconds
        distance_km:
          anyOf:
          - type: number
          - type: 'null'
          title: Distance Km
        distance_meters:
          anyOf:
          - type: number
          - type: 'null'
          title: Distance Meters
        average_heart_rate:
          anyOf:
          - type: integer
          - type: 'null'
          title: Average Heart Rate
        calories:
          anyOf:
          - type: integer
          - type: 'null'
          title: Calories
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        started_at:
          type: string
          format: date-time
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - user_id
      - local_date
      - activity
      - activity_type
      - duration_minutes
      - duration_seconds
      - started_at
      - version
      - created_at
      - updated_at
      title: CardioSessionOut
    CardioSessionPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CardioSessionOut'
          type: array
          title: Items
        cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: CardioSessionPage
    CardioSessionUpsert:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        client_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Id
        source:
          type: string
          maxLength: 32
          title: Source
          default: android
        source_device:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Source Device
        client_version:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Client Version
        local_date:
          type: string
          format: date
          title: Local Date
        activity:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Activity
        activity_type:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Activity Type
        duration_minutes:
          anyOf:
          - type: integer
            maximum: 10080.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Duration Minutes
        duration_seconds:
          anyOf:
          - type: integer
            maximum: 604800.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Duration Seconds
        distance_km:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Distance Km
        distance_meters:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Distance Meters
        average_heart_rate:
          anyOf:
          - type: integer
            maximum: 300.0
            minimum: 20.0
          - type: 'null'
          title: Average Heart Rate
        calories:
          anyOf:
          - type: integer
            maximum: 100000.0
            minimum: 0.0
          - type: 'null'
          title: Calories
        notes:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Notes
        started_at:
          type: string
          format: date-time
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        expected_version:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Expected Version
      type: object
      required:
      - id
      - local_date
      - started_at
      title: CardioSessionUpsert
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          maxLength: 254
          minLength: 3
          title: Email
        password:
          type: string
          maxLength: 1024
          minLength: 1
          title: Password
        device_name:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Device Name
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LogoutRequest:
      properties:
        refresh_token:
          anyOf:
          - type: string
            maxLength: 2048
            minLength: 32
          - type: 'null'
          title: Refresh Token
      additionalProperties: false
      type: object
      title: LogoutRequest
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          maxLength: 2048
          minLength: 32
          title: Refresh Token
      additionalProperties: false
      type: object
      required:
      - refresh_token
      title: RefreshRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: Bearer
        expires_in:
          type: integer
          title: Expires In
        expires_at:
          type: integer
          title: Expires At
      type: object
      required:
      - access_token
      - expires_in
      - expires_at
      title: TokenResponse
    UserResponse:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        display_name:
          type: string
          title: Display Name
        timezone:
          type: string
          title: Timezone
        weight_unit:
          type: string
          title: Weight Unit
          default: KG
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
          default: false
        roles:
          items:
            type: string
          type: array
          title: Roles
        permissions:
          items:
            type: string
          type: array
          title: Permissions
        version:
          type: integer
          title: Version
          default: 1
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - id
      - email
      - display_name
      - timezone
      - is_active
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer